chore: separated and modulatized  the UI components
