+++ app/page.tsx
@@ -1,95 +1,19 @@
 "use client";
 
-import Image from "next/image";
-import { useCallback, useEffect, useState, useRef, useMemo } from "react";
 import HenryWordmark from "@/assets/henry-wordmark";
+import Header from "@/components/Header";
+import ProductGrid from "@/components/ProductGrid";
+import ProductModal from "@/components/ProductModal";
+import SearchBar from "@/components/SearchBar";
+import type { Product, ProductDetails } from "@/lib/types";
 import { getValidImageUrl } from "@/lib/utils";
-
-interface Product {
-  id: string;
-  name: string;
-  price: number;
-  imageUrl: string;
-  productLink: string;
-  source: string;
-}
-
-interface ProductDetails {
-  productResults: {
-    title: string;
-    brand: string;
-    reviews: number;
-    rating: number;
-    image?: string;
-    thumbnails?: string[];
-    stores: Array<{
-      name: string;
-      link: string;
-      price: string;
-      shipping: string;
-      total: string;
-    }>;
-    variants: Array<{
-      title: string;
-      items: Array<{
-        name: string;
-        selected?: boolean;
-        available?: boolean;
-      }>;
-    }>;
-  };
-  relatedSearches: Array<{
-    query: string;
-    image: string;
-    link: string;
-  }>;
-}
+import {
+  buildDefaultVariantSelections,
+  findVariantSelection,
+} from "@/lib/variants";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 
 type ViewMode = "desktop" | "mobile";
-
-const buildDefaultVariantSelections = (
-  details: ProductDetails
-): Record<string, string> => {
-  const variants = details.productResults.variants;
-  if (!variants || variants.length === 0) {
-    return {};
-  }
-
-  return variants.reduce<Record<string, string>>((acc, variant) => {
-    const selectedItem =
-      variant.items.find((item) => item.selected && item.available !== false) ||
-      variant.items.find((item) => item.available !== false) ||
-      variant.items[0];
-
-    if (selectedItem) {
-      acc[variant.title] = selectedItem.name;
-    }
-
-    return acc;
-  }, {});
-};
-
-const getVariantPriority = (title: string) => {
-  const normalized = title.toLowerCase();
-  if (normalized.includes("size")) return 2;
-  if (normalized.includes("color")) return 1;
-  return 0;
-};
-
-const findVariantSelection = (
-  variants: ProductDetails["productResults"]["variants"],
-  selected: Record<string, string>,
-  keyword: string
-) => {
-  const match = variants.find((variant) =>
-    variant.title.toLowerCase().includes(keyword)
-  );
-  if (!match) return null;
-  return {
-    title: match.title,
-    value: selected[match.title],
-  };
-};
 
 export default function Home() {
   const [searchQuery, setSearchQuery] = useState("");
@@ -108,7 +32,7 @@ export default function Home() {
   const [checkoutIframeUrl, setCheckoutIframeUrl] = useState<string | null>(
     null
   );
-  const [iframeLoading, setIframeLoading] = useState(true);
+
   const [showProductModal, setShowProductModal] = useState(false);
   const [placeholderIndex, setPlaceholderIndex] = useState(0);
   const [viewMode, setViewMode] = useState<ViewMode>("desktop");
@@ -129,36 +53,10 @@ export default function Home() {
     "Running gear",
   ];
 
-  // Generate user ID only on client side
-  useEffect(() => {
-    setUserId(`user_${Math.random().toString(36).substring(7)}`);
-  }, []);
-
-  // Detect mobile device
-  useEffect(() => {
-    const checkMobile = () => {
-      setIsMobile(window.innerWidth < 768);
-    };
-
-    checkMobile();
-    window.addEventListener("resize", checkMobile);
-
-    return () => window.removeEventListener("resize", checkMobile);
-  }, []);
-
-  // Rotate search placeholders
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setPlaceholderIndex((prev) => (prev + 1) % placeholders.length);
-    }, 3000);
-    return () => clearInterval(interval);
-  }, []);
-
   // Handle closing the iframe
   const handleCloseIframe = useCallback(() => {
     setShowCheckoutIframe(false);
     setCheckoutIframeUrl(null);
-    setIframeLoading(true);
   }, []);
 
   // Close product modal
@@ -203,16 +101,6 @@ export default function Home() {
     }, {});
   }, [productDetails, selectedVariants]);
 
-  const sortedVariants = useMemo(() => {
-    const variants = productDetails?.productResults?.variants;
-    if (!variants || variants.length === 0) {
-      return [];
-    }
-    return [...variants].sort(
-      (a, b) => getVariantPriority(b.title) - getVariantPriority(a.title)
-    );
-  }, [productDetails]);
-
   const primarySelections = useMemo(() => {
     const variants = productDetails?.productResults?.variants ?? [];
     return {
@@ -220,31 +108,6 @@ export default function Home() {
       color: findVariantSelection(variants, selectedVariants, "color"),
     };
   }, [productDetails, selectedVariants]);
-
-  // Listen for iframe completion messages
-  useEffect(() => {
-    if (!showCheckoutIframe) return;
-
-    const handleMessage = (event: MessageEvent) => {
-      // Check if message is from iframe with completion status
-      if (event.data) {
-        const { action, orderId, status } = event.data;
-
-        // Check for checkout closed action from Done button
-        if (action === "checkoutClosed") {
-          console.log("Checkout closed", { orderId, status });
-          handleCloseIframe();
-        }
-      }
-    };
-
-    window.addEventListener("message", handleMessage);
-
-    // Cleanup listener
-    return () => {
-      window.removeEventListener("message", handleMessage);
-    };
-  }, [showCheckoutIframe, handleCloseIframe]);
 
   // Search for products
   const searchProducts = async () => {
@@ -406,694 +269,119 @@ export default function Home() {
     }
   };
 
+  // Generate user ID only on client side
+  useEffect(() => {
+    setUserId(`user_${Math.random().toString(36).substring(7)}`);
+  }, []);
+
+  // Detect mobile device
+  useEffect(() => {
+    const checkMobile = () => {
+      setIsMobile(window.innerWidth < 768);
+    };
+
+    checkMobile();
+    window.addEventListener("resize", checkMobile);
+
+    return () => window.removeEventListener("resize", checkMobile);
+  }, []);
+
+  // Rotate search placeholders
+  useEffect(() => {
+    const interval = setInterval(() => {
+      setPlaceholderIndex((prev) => (prev + 1) % placeholders.length);
+    }, 3000);
+    return () => clearInterval(interval);
+  }, []);
+
+  // Listen for iframe completion messages
+  useEffect(() => {
+    if (!showCheckoutIframe) return;
+
+    const handleMessage = (event: MessageEvent) => {
+      // Check if message is from iframe with completion status
+      if (event.data) {
+        const { action, orderId, status } = event.data;
+
+        // Check for checkout closed action from Done button
+        if (action === "checkoutClosed") {
+          console.log("Checkout closed", { orderId, status });
+          handleCloseIframe();
+        }
+      }
+    };
+
+    window.addEventListener("message", handleMessage);
+
+    // Cleanup listener
+    return () => {
+      window.removeEventListener("message", handleMessage);
+    };
+  }, [showCheckoutIframe, handleCloseIframe]);
+
   return (
     <main className="min-h-screen bg-gray-50">
       {heroView ? (
         <div className="min-h-screen flex flex-col">
-          {/* Top bar */}
-          <div className="bg-white py-4">
-            <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-              <div className="flex justify-end items-center space-x-4">
-                <div className="text-xs text-gray-500">{userId}</div>
-              </div>
-            </div>
-          </div>
-
-          {/* Centered content */}
+          <Header userId={userId} />
           <div className="flex-1 flex flex-col items-center justify-center px-4 -mt-32">
-            {/* Henry Labs Logo */}
             <HenryWordmark className="h-16 text-[#44c57e] mb-4" />
-
-            {/* Circular Search Bar */}
             <div className="w-full max-w-2xl">
-              <div className="relative">
-                <div className="absolute inset-y-0 left-0 pl-5 flex items-center pointer-events-none">
-                  <svg
-                    className="w-5 h-5 text-gray-400"
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
-                    />
-                  </svg>
-                </div>
-                <input
-                  ref={searchInputRef}
-                  type="text"
-                  value={searchQuery}
-                  onChange={(e) => setSearchQuery(e.target.value)}
-                  onKeyDown={(e) => e.key === "Enter" && searchProducts()}
-                  placeholder={placeholders[placeholderIndex]}
-                  className="w-full pl-12 pr-20 py-4 text-lg bg-white border-2 border-gray-200 rounded-full focus:outline-none focus:border-[#44c57e] focus:ring-4 focus:ring-[#44c57e]/20 transition-all duration-200 shadow-lg"
-                />
-                <button
-                  onClick={searchProducts}
-                  disabled={loading}
-                  className="absolute right-2 top-1/2 transform -translate-y-1/2 px-4 py-2 bg-[#44c57e] text-white rounded-full hover:bg-[#3aaa6a] disabled:opacity-50 transition-colors duration-200"
-                >
-                  {loading ? (
-                    <svg
-                      className="animate-spin h-5 w-5"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                    >
-                      <circle
-                        className="opacity-25"
-                        cx="12"
-                        cy="12"
-                        r="10"
-                        stroke="currentColor"
-                        strokeWidth="4"
-                      ></circle>
-                      <path
-                        className="opacity-75"
-                        fill="currentColor"
-                        d="M12 2a10 10 0 00-10 10h4a6 6 0 016-6V2zM2 12a10 10 0 0010 10v-4a6 6 0 01-6-6H2z"
-                      ></path>
-                    </svg>
-                  ) : (
-                    "Search"
-                  )}
-                </button>
-              </div>
+              <SearchBar
+                value={searchQuery}
+                onChange={setSearchQuery}
+                onSubmit={searchProducts}
+                loading={loading}
+                placeholder={placeholders[placeholderIndex]}
+                inputRef={searchInputRef}
+              />
             </div>
           </div>
         </div>
       ) : (
         <>
-          {/* Normal Header */}
-          <header className="bg-white shadow-sm sticky top-0 z-40">
-            <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-              <div className="flex items-center justify-between h-16">
-                {/* Logo */}
-                <div className="flex items-center">
-                  <HenryWordmark className="h-8 text-[#44c57e]" />
-                </div>
-
-                {/* Checkout Options */}
-                <div className="flex items-center space-x-4">
-                  <div className="text-xs text-gray-500">{userId}</div>
-                </div>
-              </div>
-            </div>
-          </header>
-
+          <Header userId={userId} />
           <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-12">
-            {/* Circular Search Bar */}
             <div className="mb-12 flex justify-center">
-              <div className="relative w-full max-w-2xl">
-                <div className="absolute inset-y-0 left-0 pl-5 flex items-center pointer-events-none">
-                  <svg
-                    className="w-5 h-5 text-gray-400"
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
-                    />
-                  </svg>
-                </div>
-                <input
-                  ref={searchInputRef}
-                  type="text"
-                  value={searchQuery}
-                  onChange={(e) => setSearchQuery(e.target.value)}
-                  onKeyDown={(e) => e.key === "Enter" && searchProducts()}
-                  placeholder={placeholders[placeholderIndex]}
-                  className="w-full pl-12 pr-20 py-4 text-lg bg-white border-2 border-gray-200 rounded-full focus:outline-none focus:border-[#44c57e] focus:ring-4 focus:ring-[#44c57e]/20 transition-all duration-200 shadow-lg"
-                />
-                <button
-                  onClick={searchProducts}
-                  disabled={loading}
-                  className="absolute right-2 top-1/2 transform -translate-y-1/2 px-4 py-2 bg-[#44c57e] text-white rounded-full hover:bg-[#3aaa6a] disabled:opacity-50 transition-colors duration-200"
-                >
-                  {loading ? (
-                    <svg
-                      className="animate-spin h-5 w-5"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                    >
-                      <circle
-                        className="opacity-25"
-                        cx="12"
-                        cy="12"
-                        r="10"
-                        stroke="currentColor"
-                        strokeWidth="4"
-                      ></circle>
-                      <path
-                        className="opacity-75"
-                        fill="currentColor"
-                        d="M12 2a10 10 0 00-10 10h4a6 6 0 016-6V2zM2 12a10 10 0 0010 10v-4a6 6 0 01-6-6H2z"
-                      ></path>
-                    </svg>
-                  ) : (
-                    "Search"
-                  )}
-                </button>
-              </div>
+              <SearchBar
+                value={searchQuery}
+                onChange={setSearchQuery}
+                onSubmit={searchProducts}
+                loading={loading}
+                placeholder={placeholders[placeholderIndex]}
+                inputRef={searchInputRef}
+                className="w-full max-w-2xl"
+              />
             </div>
-
-            {/* Products Grid */}
-            {loading ? (
-              <div className="product-grid">
-                {[...Array(15)].map((_, index) => (
-                  <div
-                    key={`skeleton-${index}`}
-                    className="bg-white rounded-xl shadow-sm overflow-hidden"
-                  >
-                    <div className="skeleton aspect-square" />
-                    <div className="p-4 space-y-3">
-                      <div className="skeleton h-3 w-16 rounded" />
-                      <div className="space-y-2">
-                        <div className="skeleton h-3 w-full rounded" />
-                        <div className="skeleton h-3 w-3/4 rounded" />
-                      </div>
-                      <div className="pt-2">
-                        <div className="skeleton h-5 w-20 rounded" />
-                      </div>
-                    </div>
-                  </div>
-                ))}
-              </div>
-            ) : products.length === 0 ? (
-              <div className="text-gray-500 text-center py-16">
-                <svg
-                  className="mx-auto h-24 w-24 text-gray-300 mb-4"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={1}
-                    d="M20 7l-8-4-8 4m16 0l-8 4m8-4v10l-8 4m0-10L4 7m8 4v10M4 7v10l8 4"
-                  />
-                </svg>
-                <p className="text-lg font-medium">No products found</p>
-                <p className="text-sm mt-2">Try searching for something!</p>
-              </div>
-            ) : (
-              <div className="product-grid">
-                {products.map((product) => (
-                  <div
-                    key={product.id}
-                    onClick={() => getProductDetails(product)}
-                    className="bg-white rounded-lg shadow-sm hover:shadow-xl transition-all duration-200 cursor-pointer overflow-hidden group"
-                  >
-                    {product.imageUrl && (
-                      <div className="relative aspect-square overflow-hidden bg-gray-100">
-                        <Image
-                          src={product.imageUrl}
-                          alt={product.name}
-                          fill
-                          className="object-cover group-hover:scale-110 transition-transform duration-300"
-                          unoptimized
-                        />
-                      </div>
-                    )}
-                    <div className="p-4 flex flex-col h-32">
-                      <p className="text-xs text-gray-500 mb-1 uppercase tracking-wide">
-                        {product.source}
-                      </p>
-                      <h3 className="text-sm text-gray-900 line-clamp-2 mb-auto">
-                        {product.name}
-                      </h3>
-                      <div className="pt-2">
-                        <p className="text-base font-bold text-black">
-                          ${product.price.toFixed(2)}
-                        </p>
-                      </div>
-                    </div>
-                  </div>
-                ))}
-              </div>
-            )}
+            <ProductGrid
+              products={products}
+              loading={loading}
+              onSelect={getProductDetails}
+            />
           </div>
 
           {/* Product Details Modal */}
-          {showProductModal && selectedProduct && (
-            <div className="fixed inset-0 z-50 flex items-center justify-center p-4">
-              {/* Overlay */}
-              <div
-                className="absolute inset-0 bg-black bg-opacity-50 modal-overlay"
-                onClick={() => !showCheckoutIframe && handleCloseProductModal()}
-              />
-
-              {/* Modal Content */}
-              <div
-                className={`relative bg-white rounded-2xl shadow-2xl modal-content overflow-hidden transition-all duration-300 ${
-                  viewMode === "desktop"
-                    ? "w-[min(1512px,90vw)] h-[min(982px,90vh)]"
-                    : "w-[430px] h-[932px]"
-                }`}
-              >
-                {!showCheckoutIframe ? (
-                  <>
-                    {/* Modal Header */}
-                    <div className="flex items-center justify-between p-4 border-b">
-                      <h2 className="text-xl font-bold">Product Details</h2>
-                      <div className="flex items-center gap-2">
-                        {/* View Mode Toggle */}
-                        <div className="flex bg-gray-100 rounded-lg p-1">
-                          <button
-                            onClick={() => setViewMode("desktop")}
-                            className={`p-2 rounded transition-colors ${
-                              viewMode === "desktop"
-                                ? "bg-white shadow-sm"
-                                : "hover:bg-gray-200"
-                            }`}
-                            aria-label="Desktop view"
-                            title="Desktop view"
-                          >
-                            <svg
-                              className="w-5 h-5"
-                              fill="none"
-                              stroke="currentColor"
-                              viewBox="0 0 24 24"
-                            >
-                              <path
-                                strokeLinecap="round"
-                                strokeLinejoin="round"
-                                strokeWidth={2}
-                                d="M9.75 17L9 20l-1 1h8l-1-1-.75-3M3 13h18M5 17h14a2 2 0 002-2V5a2 2 0 00-2-2H5a2 2 0 00-2 2v10a2 2 0 002 2z"
-                              />
-                            </svg>
-                          </button>
-                          <button
-                            onClick={() => setViewMode("mobile")}
-                            className={`p-2 rounded transition-colors ${
-                              viewMode === "mobile"
-                                ? "bg-white shadow-sm"
-                                : "hover:bg-gray-200"
-                            }`}
-                            aria-label="Mobile view"
-                            title="Mobile view"
-                          >
-                            <svg
-                              className="w-5 h-5"
-                              fill="none"
-                              stroke="currentColor"
-                              viewBox="0 0 24 24"
-                            >
-                              <path
-                                strokeLinecap="round"
-                                strokeLinejoin="round"
-                                strokeWidth={2}
-                                d="M12 18h.01M8 21h8a2 2 0 002-2V5a2 2 0 00-2-2H8a2 2 0 00-2 2v14a2 2 0 002 2z"
-                              />
-                            </svg>
-                          </button>
-                        </div>
-                        <button
-                          onClick={handleCloseProductModal}
-                          className="p-2 hover:bg-gray-100 rounded-full transition-colors"
-                          aria-label="Close"
-                        >
-                          <svg
-                            className="w-6 h-6"
-                            fill="none"
-                            stroke="currentColor"
-                            viewBox="0 0 24 24"
-                          >
-                            <path
-                              strokeLinecap="round"
-                              strokeLinejoin="round"
-                              strokeWidth={2}
-                              d="M6 18L18 6M6 6l12 12"
-                            />
-                          </svg>
-                        </button>
-                      </div>
-                    </div>
-
-                    {/* Modal Body */}
-                    <div
-                      className="relative overflow-y-auto"
-                      style={{ height: "calc(100% - 72px)" }}
-                    >
-                      {loading && (
-                        <div className="absolute inset-0 flex items-center justify-center bg-white z-10">
-                          <div className="flex flex-col items-center">
-                            <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-[#44c57e]"></div>
-                            <p className="mt-4 text-gray-600">
-                              Loading product details...
-                            </p>
-                          </div>
-                        </div>
-                      )}
-                      {productDetails ? (
-                        <div className="p-4">
-                          <div className="flex flex-col lg:grid lg:grid-cols-2 gap-4">
-                            {/* Product Image */}
-                            <div className="space-y-4">
-                              {/* Main Image */}
-                              <div className="relative h-80 rounded-lg overflow-hidden bg-white shadow-inner">
-                                {productDetails.productResults.thumbnails &&
-                                productDetails.productResults.thumbnails
-                                  .length > 0 ? (
-                                  <>
-                                    <div className="absolute inset-0 image-gradient-overlay z-10 pointer-events-none" />
-                                    <Image
-                                      src={
-                                        productDetails.productResults
-                                          .thumbnails![selectedThumbnailIndex]
-                                      }
-                                      alt={productDetails.productResults.title}
-                                      fill
-                                      className="object-contain p-4"
-                                      unoptimized
-                                    />
-                                  </>
-                                ) : productDetails.productResults.image ||
-                                  selectedProduct.imageUrl ||
-                                  productDetails.relatedSearches?.[0]?.image ? (
-                                  <>
-                                    <div className="absolute inset-0 image-gradient-overlay z-10 pointer-events-none" />
-                                    <Image
-                                      src={
-                                        productDetails.productResults.image ||
-                                        selectedProduct.imageUrl ||
-                                        productDetails.relatedSearches[0].image
-                                      }
-                                      alt={productDetails.productResults.title}
-                                      fill
-                                      className="object-contain p-4"
-                                      unoptimized
-                                    />
-                                  </>
-                                ) : (
-                                  <div className="flex items-center justify-center h-full bg-gray-50">
-                                    <svg
-                                      className="w-24 h-24 text-gray-400"
-                                      fill="none"
-                                      stroke="currentColor"
-                                      viewBox="0 0 24 24"
-                                    >
-                                      <path
-                                        strokeLinecap="round"
-                                        strokeLinejoin="round"
-                                        strokeWidth={1}
-                                        d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z"
-                                      />
-                                    </svg>
-                                  </div>
-                                )}
-                              </div>
-
-                              {/* Thumbnails Carousel */}
-                              {productDetails.productResults.thumbnails &&
-                                productDetails.productResults.thumbnails
-                                  .length > 1 && (
-                                  <div className="overflow-x-auto pb-2 max-w-full [&::-webkit-scrollbar]:h-1 [&::-webkit-scrollbar-thumb]:bg-gray-300 [&::-webkit-scrollbar-thumb]:rounded-full [&::-webkit-scrollbar-track]:bg-gray-100">
-                                    <div className="flex gap-2 py-1 px-0.5 min-w-min">
-                                      {productDetails.productResults.thumbnails!.map(
-                                        (thumbnail, index) => (
-                                          <button
-                                            key={index}
-                                            onClick={() =>
-                                              setSelectedThumbnailIndex(index)
-                                            }
-                                            className={`relative flex-shrink-0 w-14 h-14 rounded-md border-2 transition-all bg-white ${
-                                              selectedThumbnailIndex === index
-                                                ? "border-[#44c57e] opacity-100 shadow-md"
-                                                : "border-gray-300 opacity-80 hover:opacity-100 hover:border-gray-400"
-                                            }`}
-                                          >
-                                            <Image
-                                              src={thumbnail}
-                                              alt={`${
-                                                productDetails.productResults
-                                                  .title
-                                              } - View ${index + 1}`}
-                                              fill
-                                              className="object-contain p-1.5 rounded-sm"
-                                              unoptimized
-                                            />
-                                          </button>
-                                        )
-                                      )}
-                                    </div>
-                                  </div>
-                                )}
-                            </div>
-
-                            {/* Product Info */}
-                            <div className="space-y-3">
-                              <div>
-                                <h3 className="text-xl font-bold mb-1">
-                                  {productDetails.productResults.title}
-                                </h3>
-                                <p className="text-sm text-gray-600">
-                                  by {productDetails.productResults.brand}
-                                </p>
-                              </div>
-
-                              {/* Price */}
-                              <div className="text-2xl font-bold text-[#44c57e]">
-                                ${selectedProduct.price}
-                              </div>
-
-                              {/* Rating */}
-                              <div className="flex items-center gap-2">
-                                <div className="flex">
-                                  {[...Array(5)].map((_, i) => (
-                                    <span
-                                      key={i}
-                                      className={
-                                        i <
-                                        Math.floor(
-                                          productDetails.productResults.rating
-                                        )
-                                          ? "text-yellow-500"
-                                          : "text-gray-300"
-                                      }
-                                    >
-                                      ★
-                                    </span>
-                                  ))}
-                                </div>
-                                <span className="text-sm text-gray-600">
-                                  {productDetails.productResults.rating}/5 (
-                                  {productDetails.productResults.reviews}{" "}
-                                  reviews)
-                                </span>
-                              </div>
-
-                              {(primarySelections.size ||
-                                primarySelections.color) && (
-                                <div className="flex flex-wrap gap-4 text-sm text-gray-600">
-                                  {primarySelections.size && (
-                                    <span>
-                                      Size:{" "}
-                                      <span className="font-semibold">
-                                        {primarySelections.size.value ||
-                                          "Select a size"}
-                                      </span>
-                                    </span>
-                                  )}
-                                  {primarySelections.color && (
-                                    <span>
-                                      Color:{" "}
-                                      <span className="font-semibold">
-                                        {primarySelections.color.value ||
-                                          "Select a color"}
-                                      </span>
-                                    </span>
-                                  )}
-                                </div>
-                              )}
-
-                              {/* Variants */}
-                              {sortedVariants.map((variant) => (
-                                <div key={variant.title}>
-                                  <h4 className="font-medium mb-2">
-                                    {variant.title}:
-                                  </h4>
-                                  <div className="flex flex-wrap gap-2">
-                                    {variant.items.map((item) => {
-                                      const isAvailable =
-                                        item.available !== false;
-                                      const isSelected =
-                                        selectedVariants[variant.title] ===
-                                        item.name;
-
-                                      return (
-                                        <button
-                                          key={item.name}
-                                          type="button"
-                                          disabled={!isAvailable}
-                                          onClick={() => {
-                                            if (isAvailable) {
-                                              handleVariantSelection(
-                                                variant.title,
-                                                item.name
-                                              );
-                                            }
-                                          }}
-                                          aria-pressed={isSelected}
-                                          className={`px-4 py-2 border rounded-lg transition-colors ${
-                                            isSelected
-                                              ? "bg-[#44c57e] text-white border-[#44c57e]"
-                                              : isAvailable
-                                              ? "bg-white hover:bg-gray-50 border-gray-300"
-                                              : "bg-gray-100 text-gray-400 border-gray-200 cursor-not-allowed"
-                                          }`}
-                                        >
-                                          {item.name}
-                                        </button>
-                                      );
-                                    })}
-                                  </div>
-                                </div>
-                              ))}
-
-                              {/* Checkout Buttons */}
-                              <div className="pt-4 space-y-3">
-                                <button
-                                  onClick={buyNow}
-                                  disabled={loadingCheckout}
-                                  className="w-full py-3 bg-[#44c57e] text-white rounded-lg font-semibold hover:bg-[#3aaa6a] disabled:opacity-50 transition-colors"
-                                >
-                                  {loadingCheckout
-                                    ? loadingMessage
-                                    : "Add to Cart & Buy"}
-                                </button>
-
-                                {errorMessage && (
-                                  <div className="p-3 bg-red-50 border border-red-200 rounded-lg">
-                                    <p className="text-sm text-red-700">
-                                      {errorMessage}
-                                    </p>
-                                  </div>
-                                )}
-                              </div>
-                            </div>
-                          </div>
-                        </div>
-                      ) : !loading ? (
-                        <div className="p-6 text-center py-12 text-gray-500">
-                          Unable to load product details
-                        </div>
-                      ) : null}
-                    </div>
-                  </>
-                ) : (
-                  /* Checkout Iframe within Modal */
-                  <div className="relative h-full flex flex-col">
-                    <div className="flex items-center justify-between p-4 border-b bg-gray-50 flex-shrink-0">
-                      <h3 className="text-lg font-semibold">
-                        Complete Your Purchase
-                      </h3>
-                      <div className="flex items-center gap-2">
-                        {/* View Mode Toggle */}
-                        <div className="flex bg-gray-100 rounded-lg p-1">
-                          <button
-                            onClick={() => setViewMode("desktop")}
-                            className={`p-2 rounded transition-colors ${
-                              viewMode === "desktop"
-                                ? "bg-white shadow-sm"
-                                : "hover:bg-gray-200"
-                            }`}
-                            aria-label="Desktop view"
-                            title="Desktop view"
-                          >
-                            <svg
-                              className="w-5 h-5"
-                              fill="none"
-                              stroke="currentColor"
-                              viewBox="0 0 24 24"
-                            >
-                              <path
-                                strokeLinecap="round"
-                                strokeLinejoin="round"
-                                strokeWidth={2}
-                                d="M9.75 17L9 20l-1 1h8l-1-1-.75-3M3 13h18M5 17h14a2 2 0 002-2V5a2 2 0 00-2-2H5a2 2 0 00-2 2v10a2 2 0 002 2z"
-                              />
-                            </svg>
-                          </button>
-                          <button
-                            onClick={() => setViewMode("mobile")}
-                            className={`p-2 rounded transition-colors ${
-                              viewMode === "mobile"
-                                ? "bg-white shadow-sm"
-                                : "hover:bg-gray-200"
-                            }`}
-                            aria-label="Mobile view"
-                            title="Mobile view"
-                          >
-                            <svg
-                              className="w-5 h-5"
-                              fill="none"
-                              stroke="currentColor"
-                              viewBox="0 0 24 24"
-                            >
-                              <path
-                                strokeLinecap="round"
-                                strokeLinejoin="round"
-                                strokeWidth={2}
-                                d="M12 18h.01M8 21h8a2 2 0 002-2V5a2 2 0 00-2-2H8a2 2 0 00-2 2v14a2 2 0 002 2z"
-                              />
-                            </svg>
-                          </button>
-                        </div>
-                        <button
-                          onClick={handleCloseIframe}
-                          className="p-2 hover:bg-gray-200 rounded-full transition-colors"
-                          aria-label="Close"
-                        >
-                          <svg
-                            className="w-6 h-6"
-                            fill="none"
-                            stroke="currentColor"
-                            viewBox="0 0 24 24"
-                          >
-                            <path
-                              strokeLinecap="round"
-                              strokeLinejoin="round"
-                              strokeWidth={2}
-                              d="M6 18L18 6M6 6l12 12"
-                            />
-                          </svg>
-                        </button>
-                      </div>
-                    </div>
-
-                    <div className="flex-1 relative">
-                      {iframeLoading && (
-                        <div className="absolute inset-0 flex items-center justify-center bg-white z-10">
-                          <div className="flex flex-col items-center">
-                            <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-[#44c57e]"></div>
-                            <p className="mt-4 text-gray-600">
-                              Loading checkout...
-                            </p>
-                          </div>
-                        </div>
-                      )}
-
-                      <iframe
-                        src={checkoutIframeUrl ?? undefined}
-                        className="w-full h-full"
-                        title="Checkout"
-                        allow="payment"
-                        onLoad={() => setIframeLoading(false)}
-                      />
-                    </div>
-                  </div>
-                )}
-              </div>
-            </div>
-          )}
+          <ProductModal
+            isOpen={showProductModal && !!selectedProduct}
+            onClose={handleCloseProductModal}
+            viewMode={viewMode}
+            setViewMode={setViewMode}
+            loading={loading}
+            loadingCheckout={loadingCheckout}
+            loadingMessage={loadingMessage}
+            productDetails={productDetails}
+            selectedProduct={selectedProduct}
+            selectedThumbnailIndex={selectedThumbnailIndex}
+            setSelectedThumbnailIndex={setSelectedThumbnailIndex}
+            selectedVariants={selectedVariants}
+            onSelectVariant={handleVariantSelection}
+            primarySelections={primarySelections as any}
+            buyNow={buyNow}
+            errorMessage={errorMessage}
+            showCheckoutIframe={showCheckoutIframe}
+            checkoutIframeUrl={checkoutIframeUrl}
+            onCloseIframe={handleCloseIframe}
+          />
         </>
       )}
     </main>
