+++ lib/types.ts
@@ -1,0 +1,40 @@
+export interface Product {
+  id: string;
+  name: string;
+  price: number;
+  imageUrl: string;
+  productLink: string;
+  source: string;
+}
+
+export interface ProductDetails {
+  productResults: {
+    title: string;
+    brand: string;
+    reviews: number;
+    rating: number;
+    image?: string;
+    thumbnails?: string[];
+    stores: Array<{
+      name: string;
+      link: string;
+      price: string;
+      shipping: string;
+      total: string;
+    }>;
+    variants: Array<{
+      title: string;
+      items: Array<{
+        name: string;
+        selected?: boolean;
+        available?: boolean;
+      }>;
+    }>;
+  };
+  relatedSearches: Array<{
+    query: string;
+    image: string;
+    link: string;
+  }>;
+}
+
